remove translation on untranslated dynamic type
