+++ src/app/js/admin/enrichment/EnrichmentCatalog.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+
+import enrichers from '../../../custom/enrichers/enrichers-catalog.json';
+import Catalog from '../../lib/components/Catalog';
+
+const translatePrefix = 'ws';
+
+type EnrichmentCatalogProps = {
+    isOpen: boolean;
+    handleClose: () => void;
+    onChange: () => void;
+    selectedWebServiceUrl: string;
+};
+
+export const EnrichmentCatalog = ({
+    isOpen,
+    handleClose,
+    onChange,
+    selectedWebServiceUrl,
+}: EnrichmentCatalogProps) => (
+    <Catalog
+        {...{
+            isOpen,
+            handleClose,
+            onChange,
+            selectedWebServiceUrl,
+            enrichers,
+            translatePrefix,
+        }}
+    />
+);
+
+export default EnrichmentCatalog;
